// src/types/models/CalendarVO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gym Manager Swagger
 * Gym Manager REST API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CalendarVO
 */
export interface CalendarVO {
    /**
     * 일정 내용
     * @type {string}
     * @memberof CalendarVO
     */
    clContent?: string;
    /**
     * 일정 메모
     * @type {string}
     * @memberof CalendarVO
     */
    clMemo?: string;
    /**
     * 일정관리 시퀀스
     * @type {number}
     * @memberof CalendarVO
     */
    clSeq?: number;
    /**
     * 일정 타입
     * @type {string}
     * @memberof CalendarVO
     */
    clType?: string;
    /**
     * 작성 일자
     * @type {Date}
     * @memberof CalendarVO
     */
    regDate?: Date;
    /**
     * 작성자 ID
     * @type {string}
     * @memberof CalendarVO
     */
    regId?: string;
    /**
     * 사업장 시퀀스
     * @type {number}
     * @memberof CalendarVO
     */
    wpSeq?: number;
}

/**
 * Check if a given object implements the CalendarVO interface.
 */
export function instanceOfCalendarVO(value: object): value is CalendarVO {
    return true;
}

export function CalendarVOFromJSON(json: any): CalendarVO {
    return CalendarVOFromJSONTyped(json, false);
}

export function CalendarVOFromJSONTyped(json: any, ignoreDiscriminator: boolean): CalendarVO {
    if (json == null) {
        return json;
    }
    return {
        
        'clContent': json['clContent'] == null ? undefined : json['clContent'],
        'clMemo': json['clMemo'] == null ? undefined : json['clMemo'],
        'clSeq': json['clSeq'] == null ? undefined : json['clSeq'],
        'clType': json['clType'] == null ? undefined : json['clType'],
        'regDate': json['regDate'] == null ? undefined : (new Date(json['regDate'])),
        'regId': json['regId'] == null ? undefined : json['regId'],
        'wpSeq': json['wpSeq'] == null ? undefined : json['wpSeq'],
    };
}

export function CalendarVOToJSON(json: any): CalendarVO {
    return CalendarVOToJSONTyped(json, false);
}

export function CalendarVOToJSONTyped(value?: CalendarVO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'clContent': value['clContent'],
        'clMemo': value['clMemo'],
        'clSeq': value['clSeq'],
        'clType': value['clType'],
        'regDate': value['regDate'] == null ? undefined : ((value['regDate']).toISOString()),
        'regId': value['regId'],
        'wpSeq': value['wpSeq'],
    };
}